html/semantics/embedded-content/media-elements/loading-the-media-resource/stable-state-dialogs.html

<!doctype html>
<title>provide a stable state in dialogs</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
['alert', 'confirm', 'prompt'].forEach(function(dialog) {
  test(function() {
    v = document.createElement('video');
    assert_equals(v.networkState, v.NETWORK_EMPTY, 'networkState before load()');
    v.load();
    assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState after load()');
    window[dialog]('dismiss this dialog');
    assert_equals(v.networkState, v.NETWORK_EMPTY, 'networkState after alert()');
  }, dialog + '() should provide a stable state');
});
</script>
